requirements.txt
streamlit
langchain
transformers
datasets
python-dotenv
torch
keras
tensorflow
chromadb
wikipedia-api
aiohttp  # for making API calls concurrently
argparse  # for running script from command line
asyncio  # for running API calls concurrently
tiktoken  # for counting tokens
